+++ CHANGELOG.md
@@ -16,7 +16,8 @@
 <!-- Your comment below this -->
 
 - Bitbucket Cloud: Fix bug when Danger updating inline comment with summary comment. - [@hellocore]
-  <!-- Your comment above this -->
+
+<!-- Your comment above this -->
 
 # 10.4.2